// src/modules/transaction/transaction.enum.ts
export enum TransactionCategory {
  DEPOSIT = 'deposit',
  WITHDRAWAL = 'withdrawal',
  TRANSFER_IN = 'transfer in',
  TRANSFER_OUT = 'transfer out',
  TRANSFER_REVERSED = 'transfer reversed',
  REFERRAL = 'referral',
  TRADE = 'trade',
  INVESTMENT = 'investment',
  ITEM_CREATED = 'item created',
  ITEM_BOUGHT = 'item bought',
  ITEM_SOLD = 'item sold',
}